requirements.txt
openpyxl>=3.1.0
openai>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
unstructured==0.14.3
networkx
PyYAML